Restructure to allow protocols to be shared between controllers and widgets
