+++ src/sunflare/view/__init__.py
@@ -1,3 +1,3 @@
-from .protocol import WidgetProtocol
+from ._protocols import WidgetProtocol
 
 __all__ = ["WidgetProtocol"]
