+++ src/sunflare/virtual/__init__.py
@@ -1,4 +1,5 @@
 from ._bus import Publisher, Signal, Subscriber, VirtualBus, decode, encode, slot
+from ._protocols import HasConnection, HasRegistration, HasShutdown
 
 __all__ = [
     "VirtualBus",
@@ -6,6 +7,9 @@ __all__ = [
     "Signal",
     "Publisher",
     "Subscriber",
+    "HasConnection",
+    "HasRegistration",
+    "HasShutdown",
     "encode",
     "decode",
 ]
